Update README.md
This includes my favorite series and movies of all time because these movies/series had played a big role in my life for it is the reason why I am better person now. These movies had given me a lesson that I have really applied in my life as of the moment.